feat: wip router refactor
