+++ composer.json
@@ -3,7 +3,8 @@
     "autoload": {
         "psr-4": {
             "App\\Controller\\": "src/Controller",
-            "App\\Model\\": "src/Model"
+            "App\\Model\\": "src/Model",
+            "App\\View\\": "public/View"
         }
     },
     "require": {
