+++ index.php
@@ -2,27 +2,47 @@
 
 require_once 'vendor/autoload.php';
 session_start();
+
+use App\Controller\AuthenticationController;
+
+$router = new AltoRouter();
+$auth = new AuthenticationController();
+
+$router->setBasePath('/my-little-mvc');
+
+$router->map('GET', '/', function () {
+    require 'public/View/home.php';
+});
+
+$router->map('GET', '/shop', function () {
+    require 'public/View/shop.php';
+});
+
+$router->map('GET', '/login', function () {
+    require 'public/View/login.php';
+}, 'login');
+
+$router->map('POST', '/login', function () use ($auth) {
+    $reg = $auth->login($_POST['email'], $_POST['password']);
+    $message = $reg;
+    require 'public/View/login.php';
+}, 'login_submit');
+
+$router->map('GET', '/blogs', function () {
+    echo 'Blogs';
+});
+
+$match = $router->match();
+
+if ($match) {
+    call_user_func_array($match['target'], $match['params']);
+} else {
+    require_once 'public/View/404.php';
+}
+
 ?>
 
-<!doctype html>
-<html lang="fr">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <script src="https://cdn.tailwindcss.com"></script>
-    <title>Shop - Home</title>
-</head>
-<body>
-    <?php require_once 'import/header.php'; ?>
-    <main>
-        <section class="w-screen h-screen flex justify-center items-center">
-            <h1 class="text-6xl font-semibold text-[#7B41F9]">Bienvenue sur notre site</h1>
-        </section>
-    </main>
-</body>
-</html>
+
