+++ public/View/404.php
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="fr">
+<head>
+    <title>404 Not Found</title>
+</head>
+<body>
+<h1>404 Not Found</h1>
+<p>The requested page could not be found.</p>
+</body>
+</html>
