+++ public/View/home.php
@@ -1,0 +1,19 @@
+<!doctype html>
+<html lang="fr">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <script src="https://cdn.tailwindcss.com"></script>
+    <title>Shop - Home</title>
+</head>
+<body>
+    <?php require_once 'import/header.php';?>
+    <main>
+        <section class="w-screen h-screen flex justify-center items-center">
+            <h1 class="text-6xl font-semibold text-[#7B41F9]">Bienvenue sur notre site</h1>
+        </section>
+    </main>
+</body>
+</html>
